pyproject.toml
[project]
name = "auth-service"
version = "0.4.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
    "boto3>=1.38.22",
    "boto3-stubs[essential]>=1.38.22",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "mangum>=0.19.0",
    "pendulum>=3.1.0",
    "pydantic[email]>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "ujson>=5.10.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "aws-lambda-powertools[validation]>=3.13.0",
    "bandit>=1.8.3",
    "moto>=5.0.28",
    "mypy>=1.15.0",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.6.1",
    "respx>=0.22.0",
    "ruff>=0.14.7",
    "time-machine>=2.16.0",
]
